+++ packages/framework/src/common/abstract.window.ts
@@ -1,6 +1,6 @@
-import { EventEmitter } from 'events';
-import type { BrowserWindow } from 'electron';
-import type { ILogger } from '@electron-boot/logger';
+import { EventEmitter } from 'events'
+import type { BrowserWindow } from 'electron'
+import type { ILogger } from '@electron-boot/logger'
 
 /**
  * AbstractWindow 是一个抽象类，继承自 EventEmitter，用于表示应用程序中的窗口对象。
@@ -16,25 +16,46 @@ import type { ILogger } from '@electron-boot/logger';
  *   - close(): 关闭当前窗口。
  */
 export abstract class AbstractWindow extends EventEmitter {
-  protected logger: ILogger | null = null;
-  protected _id: number | null = null;
-  protected _win: BrowserWindow | null = null;
+  protected logger: ILogger | null = null
+  protected _id: number | null = null
+  protected _win: BrowserWindow | null = null
+  protected ready: null | Promise<boolean> = null
+  protected async createWindow(): Promise<void> {
+    throw new Error('Method not implemented.')
+  }
 
   send(channel: string, ...args: any[]): void {
     if (this._win) {
       if (this._win.isDestroyed() || this._win.webContents.isDestroyed()) {
-        this.logger?.warn(`Sending IPC message to channel '${channel}' for window that is destroyed`);
-        return;
+        this.logger?.warn(`Sending IPC message to channel '${channel}' for window that is destroyed`)
+        return
       }
       try {
-        this._win.webContents.send(channel, ...args);
+        this._win.webContents.send(channel, ...args)
       } catch (error) {
-        this.logger?.warn(`Error sending IPC message to channel '${channel}' of window ${this._id}: ${error}`);
+        this.logger?.warn(`Error sending IPC message to channel '${channel}' of window ${this._id}: ${error}`)
       }
     }
   }
+  /**
+   * 显示窗口
+   */
+  async show(): Promise<void> {
+    // 创建窗口
+    await this.createWindow()
+    // 显示窗口
+    if (this._win) {
+      // 等待窗口创建完毕
+      if (this.ready) await this.ready
+      // 显示窗口
+      this._win?.show()
+    }
+  }
 
+  hide(): void {
+    this._win?.hide()
+  }
   close(): void {
-    this._win?.close();
+    this._win?.close()
   }
 }